upload_embeddings_to_weaviate/requirements.txt
google-cloud-bigquery
pandas
weaviate-client
db-dtypes
